move definitions to own files
