+++ src/routes/me/definitions.ts
@@ -1,0 +1,73 @@
+import {
+  ApiEndpointDefinition,
+  ApiCallDefinition,
+  ApiResponseDefinition,
+  ApiEndpointName,
+  ApiEndpointUri
+} from "../ApiEndpoint";
+import {
+  OperationApiResponseBody,
+  MessageApiResponseBody
+} from "../ApiEndpoint/ApiResponse";
+import { HttpVerb } from "../../helpers/http";
+import { ApiLocationDefinition } from "../ApiEndpoint/ApiCall";
+
+/*
+ * TODO: find a better way to narrow the types precisely than this repetitive type declaration
+ */
+export type SendInviteApiLocation = ApiLocationDefinition<
+  ApiEndpointUri.SEND_INVITE,
+  HttpVerb.POST,
+  true
+>;
+export const sendInviteApiLocation: SendInviteApiLocation = {
+  uri: ApiEndpointUri.SEND_INVITE,
+  method: HttpVerb.POST,
+  authenticated: true
+};
+export type SendInviteApiCall = ApiCallDefinition<
+  SendInviteApiLocation["uri"],
+  SendInviteApiLocation["method"],
+  SendInviteApiLocation["authenticated"],
+  void,
+  { inviteEmail: string }
+>;
+export type SendInviteApiResponse = ApiResponseDefinition<
+  201,
+  MessageApiResponseBody
+>;
+export type SendInviteApiEndpoint = ApiEndpointDefinition<
+  ApiEndpointName.SEND_INVITE,
+  SendInviteApiCall,
+  SendInviteApiResponse
+>;
+
+/*
+ * TODO: find a better way to narrow the types precisely than this repetitive type declaration
+ */
+export type MintApiLocation = ApiLocationDefinition<
+  ApiEndpointUri.MINT,
+  HttpVerb.POST,
+  true
+>;
+export const mintApiLocation: MintApiLocation = {
+  uri: ApiEndpointUri.MINT,
+  method: HttpVerb.POST,
+  authenticated: true
+};
+export type MintApiCall = ApiCallDefinition<
+  MintApiLocation["uri"],
+  MintApiLocation["method"],
+  MintApiLocation["authenticated"],
+  void,
+  { amount: string }
+>;
+export type MintApiResponse = ApiResponseDefinition<
+  201,
+  OperationApiResponseBody
+>;
+export type MintApiEndpoint = ApiEndpointDefinition<
+  ApiEndpointName.MINT,
+  MintApiCall,
+  MintApiResponse
+>;
